*Update gradle dependencies

+++ app/src/main/java/com/vsa/filmoteca/data/repository/util/CacheManager.java
@@ -2,7 +2,6 @@ package com.vsa.filmoteca.data.repository.util;
 
 import android.content.Context;
 
-import com.vsa.filmoteca.FilmotecaApplication;
 import com.vsa.filmoteca.data.repository.ws.WSClient;
 
 import java.io.File;
@@ -23,11 +22,12 @@ public class CacheManager {
             long currentTime = System.currentTimeMillis();
             File cacheDir = new File(context.getCacheDir(), WSClient.CACHE_DIRECTORY);
             boolean isExpired;
-            for (File file : cacheDir.listFiles()) {
-                isExpired = (file.lastModified() + millis) < currentTime;
-                if (isExpired)
-                    file.delete();
-            }
+            if (cacheDir != null && cacheDir.listFiles() != null)
+                for (File file : cacheDir.listFiles()) {
+                    isExpired = (file.lastModified() + millis) < currentTime;
+                    if (isExpired)
+                        file.delete();
+                }
         }).start();
     }
